# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a retail management system server.",
        "title": "Retail Pro API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "description": "Retrieves a list of all orders associated with the logged-in user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get all orders for the authenticated user",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved orders",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.DataResponse-array_routes_order_Order"
                        }
                    },
                    "400": {
                        "description": "User ID is required",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new order for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order details",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes_order.CreateOrder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routes_order.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Retrieves a specific order by its ID for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved order",
                        "schema": {
                            "$ref": "#/definitions/routes_order.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid order ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context or not owner",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: User ID in context is not a string",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an order by its ID for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Delete an order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid order ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context or not owner",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_yantology_simple-pos_pkg_dto.DataResponse-array_routes_order_Order": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routes_order.Order"
                    }
                }
            }
        },
        "github_com_yantology_simple-pos_pkg_dto.MessageResponse": {
            "description": "Generic message response model",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "routes_order.CreateOrder": {
            "type": "object",
            "properties": {
                "product": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routes_order.Product"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "routes_order.Order": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "description": "Changed from string to int",
                    "type": "integer"
                },
                "product": {
                    "description": "Reverted back to []Product",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routes_order.Product"
                    }
                },
                "total": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "Changed from string to int",
                    "type": "integer"
                }
            }
        },
        "routes_order.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "description": "Changed from string to int",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}